Change Coordinate attributes to doubles

+++ aqmaps/src/main/java/uk/ac/ed/inf/aqmaps/Coordinates.java
@@ -1,6 +1,6 @@
 package uk.ac.ed.inf.aqmaps;
 
 public class Coordinates {
-  float lng;  // Longitude
-  float lat;  // Latitude
+  double lng;  // Longitude
+  double lat;  // Latitude
 }
